+++ docs/changelog.rst
@@ -17,11 +17,13 @@ New features
 * Plotting
 
   * :class:`aspecd.plotting.PatchProperties` for properties of patches, as used for :class:`aspecd.annotations.VerticalSpan` and :class:`aspecd.annotations.HorizontalSpan`.
+  * :class:`aspecd.plotting.AnnotationProperties` and :class:`aspecd.plotting.AnnotationTextProperties` for properties of text annotations with lines, as used for :class:`aspecd.annotations.TextWithLine`.
 
 * Plot annotations
 
   * :class:`aspecd.annotations.VerticalSpan` for adding vertical spans (rectangles) to plot(ter)s
   * :class:`aspecd.annotations.HorizontalSpan` for adding horizontal spans (rectangles) to plot(ter)s
+  * :class:`aspecd.annotations.TextWithLine` for adding text with additional lines
 
 
 Fixes